+++ src/availability/availability.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { BOMLine } from '../entities/bom-line.entity';
+import { Item } from '../entities/item.entity';
+import { Project } from '../entities/project.entity';
+import { AvailabilityController } from './availability.controller';
+import { AvailabilityService } from './availability.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Item, BOMLine, Project])],
+  controllers: [AvailabilityController],
+  providers: [AvailabilityService],
+  exports: [AvailabilityService],
+})
+export class AvailabilityModule {}